requirements.txt
selenium~=4.10.0
pytest~=7.4.0
python-dotenv~=1.0.0
